refactor(score-calc): 衣装詳細テーブルを CardDetailTable に抽出
